Structural changes: client/src -> src, vuetify added
I was having a hell of a time getting vuetify to work with our project's
structure, and I have no idea why: before, we had `client/src/`, and
having `/src/` nested in a directory seemed to cause some major trouble
in finding vuetify deps. I opted for the easy thing, which was
restructuring the project, rather than getting to the bottom of
vuetify's funk. So, don't take the current structure as the ideal
structure: we should re-consider it in order to better separate the
client and server. For now, though, this works.

Vuetify was added. It lives in its own plugin within `/src/`. I haven't
added any configurations to it, but we need to consider the following:
  - the "a la carte" method of installing deps: currently, we're pulling
    the entire lib in
  - other configuration options, including themes, icons, and so on

Webpack:
We're no longer using `client`, so it's been taken out as part of the
path. I've only commented it out rather than fully removin git. When we
reconsider the structure of the project, look at this commit and how
things were done. I'll have another commit for removing the
commented-out code. Take note of the plugins used to make vuetify work.
If we ever have our own style system, we'll want to remind ourselves of
how things were originally set up.

/scss/:
Liz added the scss directory: it was squashed alongside some others into
this commit. We'll _eventually_ use it for custom classes. Sorry about
that, Liz, if I have time later I'll rework the history for it to be
atomic, but, alas, lunch is almost over.

nvmrc:
Added, set to 10.15.0 currently. We'll look into updating that version
after we've settled on pinned versions of vue, vuetify, and the other
vue-related packages (e.g., vuex, vuelidate).