forms: hooking into FE
Forms are now being used for the RegModule (new user registration).
We're iterating over data received from the BE, the user adds to it, and
then it gets POSTed back to the BE.

A change to authorization.js was needed to fit the new data format;
similar changes were made to the RegModule component.